Core/Spells: Correct HasUnitMovementFlag (?), should fix Glyph of Momentum

+++ src/server/game/Spells/Spell.cpp
@@ -1279,20 +1279,20 @@ void Spell::SelectImplicitCasterDestTargets(SpellEffIndex effIndex, SpellImplici
     float angle = targetType.CalcDirectionAngle();
     if (targetType.GetTarget() == TARGET_DEST_DYN_MOV_FLAGS)
     {
-        if (m_caster->HasUnitMovementFlag(10)) // backward right
-            angle = -3*M_PI/4; 
-        else if (m_caster->HasUnitMovementFlag(9)) // forward right
-            angle = -M_PI/4;    
-        else if (m_caster->HasUnitMovementFlag(8)) // strafe right
-            angle = -M_PI/2;            
-        else if (m_caster->HasUnitMovementFlag(6)) // backward left
-            angle = 3*M_PI/4;
-        else if (m_caster->HasUnitMovementFlag(5)) // forward left
-            angle = M_PI/4;
-        else if (m_caster->HasUnitMovementFlag(4)) // strafe left
-            angle = M_PI/2;                             
-        else if (m_caster->HasUnitMovementFlag(2)) // backward
-            angle = M_PI;
+        if (m_caster->HasUnitMovementFlag(10)) // backward right   
+            angle = static_cast<float>(-3*M_PI/4);      
+        else if (m_caster->HasUnitMovementFlag(9)) // forward right 
+            angle = static_cast<float>(-M_PI/4);       
+        else if (m_caster->HasUnitMovementFlag(8)) // strafe right  
+            angle = static_cast<float>(-M_PI/2);       
+        else if (m_caster->HasUnitMovementFlag(6)) // backward left 
+            angle = static_cast<float>(3*M_PI/4);      
+        else if (m_caster->HasUnitMovementFlag(5)) // forward left  
+            angle = static_cast<float>(M_PI/4);       
+        else if (m_caster->HasUnitMovementFlag(4)) // strafe left   
+            angle = static_cast<float>(M_PI/2);     
+        else if (m_caster->HasUnitMovementFlag(2)) // backward      
+            angle = static_cast<float>(M_PI);                       
         else
             angle = 0.0f;
     }
